Add persistence, scheduled sweeps, and two new sources
The meter could not accumulate history: it needs roughly six samples over
half a day before it will speak, and nothing invoked by hand gets there.

- hl-core::store: append-only observation store with stable dedupe keys,
  per-source resume points, corrupt-line tolerance and retention pruning
- hl-venues::huggingface: tag saturation from one unauthenticated request
  per tag, where the span of the 100 newest artefacts is the creation rate
- hl-venues::github_search: cross-repository discovery rather than a
  fixed watchlist, on its own tighter rate bucket
- hl sweep / hl report: one pass over every source, then a ranked
  REPORT.md readable without running anything
- Hourly workflow on free public-repo runners, snapshotting state to a
  separate data branch

Competitor density widened from a count to a rate so a per-day figure can
sit in the same metric as an issue comment count; the meter only ever fits
the trend, never the level, so the two are interchangeable within a niche.

101 tests, no network required.

+++ crates/hl-core/src/types.rs
@@ -74,8 +74,11 @@ pub struct Observation {
     /// Time between an opportunity appearing and someone claiming it. The primary
     /// crowding signal: when this collapses, machines have arrived.
     pub claim_latency_ms: Option<u64>,
-    /// Visible competitors on a single opportunity.
-    pub competitors: Option<u32>,
+    /// Competitor density. Either a count on a single opportunity (comments on an
+    /// issue) or a rate (new entrants per day in a category) — both answer "how many
+    /// others are chasing this", and the meter only ever fits the trend, never the
+    /// absolute level, so the two are interchangeable within a niche.
+    pub competitors: Option<f64>,
     /// Advertised or realised reward, in cents.
     pub reward_cents: Option<u64>,
     /// Realised acceptance rate, when the venue tells us.
@@ -99,7 +102,7 @@ impl Observation {
         self.claim_latency_ms = Some(ms);
         self
     }
-    pub fn competitors(mut self, n: u32) -> Self {
+    pub fn competitors(mut self, n: f64) -> Self {
         self.competitors = Some(n);
         self
     }
